requirements.txt
Pillow
easyocr
requests